Update to include playGame function
Function hardcoded to 5 rounds via repeated if/else if "loops", which
check each round for ties and repeat the round if tie is found. Also
checks for win conditions each time in case. Understand that it's
impossible for humanWin or computerWin to trigger at any stage earlier
than round 5 (and would require 5 straight wins in a row for either) but
included in each round in order to cement learning.