Fixed 2 bugs: double-ties and double win message
1. In instances when rounds were tied twice in a row, they were counting
as a completed round, ending in final scores that did not sum 5 (eg.
human 3 - computer 1). This is fixed by a) the addition of the validRounds
variable being counted b) the ifTieResult function allowing for
recursion loop so that the code does not just continue on after running
the playRound function only once after a tie and c) the finalRounds
recursive loop repeating if the the validRounds variable != 5.

2. In the (albeit rare) event that either human or computer scored 5
wins in a row, both the specific 5-round sweep win alerts and the "win
by majority" alerts played. This is fixed by changing of the win alerts
so that they are only displayed whem human or computer score != 5.